Storage permission requested only when an app is tapped in the file chooser
Issue #181
